# apps/signalguard-api/openapi.json
{
  "components": {
    "schemas": {
      "AffectedObject": {
        "description": "A target the finding refers to. Carries enough identification for the\nLLM narrator and the UI to render the object without re-querying Graph.",
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "type": {
            "enum": [
              "policy",
              "user",
              "group",
              "role",
              "location",
              "app",
              "tenant"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "id",
          "display_name"
        ],
        "title": "AffectedObject",
        "type": "object"
      },
...
      "AuditDryRunResponse": {
        "properties": {
          "by_category": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Category",
            "type": "object"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "title": "Findings",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "findings",
          "by_category",
          "run_id"
        ],
        "title": "AuditDryRunResponse",
        "type": "object"
      },
      "AuditRunRequest": {
        "description": "Body of POST /tenants/{id}/audit/run.",
        "properties": {
          "categories": {
            "description": "Subset of audit categories to execute. Empty list rejected.",
            "items": {
              "enum": [
                "coverage",
                "rules",
                "exclusions"
              ],
              "type": "string"
            },
            "title": "Categories",
            "type": "array"
          },
          "generate_narratives": {
            "default": true,
            "description": "When true, run the LLM narrator over written findings.",
            "title": "Generate Narratives",
            "type": "boolean"
          },
          "narrative_budget_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional per-run budget cap; falls back to env CSTACK_LLM_BUDGET_USD.",
            "title": "Narrative Budget Usd"
          }
        },
        "title": "AuditRunRequest",
        "type": "object"
      },
      "AuditRunResponse": {
        "properties": {
          "by_category": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Category",
            "type": "object"
          },
          "duration_seconds": {
            "title": "Duration Seconds",
            "type": "number"
          },
          "findings_written": {
            "title": "Findings Written",
            "type": "integer"
          },
          "narrative_summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NarrativeBatchSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "findings_written",
          "by_category",
          "duration_seconds",
          "run_id"
        ],
        "title": "AuditRunResponse",
        "type": "object"
      },
...
      "Finding": {
        "description": "An immutable audit finding. Identity is computed from inputs so the\nsame finding produced twice in a row writes one row.",
        "properties": {
          "affected_objects": {
            "items": {
              "$ref": "#/components/schemas/AffectedObject"
            },
            "title": "Affected Objects",
            "type": "array"
          },
          "category": {
            "enum": [
              "coverage",
              "rule",
              "exclusion",
              "anomaly"
            ],
            "title": "Category",
            "type": "string"
          },
          "detected_at": {
            "format": "date-time",
            "title": "Detected At",
            "type": "string"
          },
          "evidence": {
            "additionalProperties": true,
            "title": "Evidence",
            "type": "object"
          },
          "first_seen_at": {
            "format": "date-time",
            "title": "First Seen At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "references": {
            "items": {
              "type": "string"
            },
            "title": "References",
            "type": "array"
          },
          "remediation_hint": {
            "title": "Remediation Hint",
            "type": "string"
          },
          "rule_id": {
            "title": "Rule Id",
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "rule_id",
          "category",
          "severity",
          "title",
          "summary",
          "affected_objects",
          "evidence",
          "remediation_hint",
          "references",
          "detected_at",
          "first_seen_at"
        ],
        "title": "Finding",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NarrativeBatchSummary": {
        "properties": {
          "cache_hits": {
            "title": "Cache Hits",
            "type": "integer"
          },
          "dollars_spent": {
            "title": "Dollars Spent",
            "type": "number"
          },
          "errored": {
            "title": "Errored",
            "type": "integer"
          },
          "generated": {
            "title": "Generated",
            "type": "integer"
          },
          "skipped_budget": {
            "title": "Skipped Budget",
            "type": "integer"
          }
        },
        "required": [
          "cache_hits",
          "generated",
          "skipped_budget",
          "errored",
          "dollars_spent"
        ],
        "title": "NarrativeBatchSummary",
        "type": "object"
      },
      "NarrativeResponse": {
        "properties": {
          "cached": {
            "title": "Cached",
            "type": "boolean"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "input_tokens": {
            "title": "Input Tokens",
            "type": "integer"
          },
          "markdown": {
            "title": "Markdown",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "output_tokens": {
            "title": "Output Tokens",
            "type": "integer"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "markdown",
          "model",
          "provider",
          "generated_at",
          "cached",
          "input_tokens",
          "output_tokens"
        ],
        "title": "NarrativeResponse",
        "type": "object"
      },
...
      "RegenerateRequest": {
        "properties": {
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "prompt_version": {
            "default": "v1",
            "title": "Prompt Version",
            "type": "string"
          }
        },
        "title": "RegenerateRequest",
        "type": "object"
      },
      "Severity": {
        "enum": [
          "INFO",
          "LOW",
          "MEDIUM",
          "HIGH",
          "CRITICAL"
        ],
        "title": "Severity",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "email": "leunis@vanlabs.dev",
      "name": "cstack maintainers"
    },
    "description": "HTTP surface over the cstack signalguard data and audit packages. See /docs for OpenAPI 3.1 and /redoc for the rendered spec.",
    "license": {
      "name": "MIT"
    },
    "title": "signalguard-api",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/tenants/{tenant_id}/audit/dry-run": {
      "post": {
        "description": "Useful for previewing what an audit run would produce against the current data without writing rows to the findings table.",
        "operationId": "dry_run_tenants__tenant_id__audit_dry_run_post",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditDryRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute findings without persisting them",
        "tags": [
          "audit"
        ]
      }
    },
    "/tenants/{tenant_id}/audit/run": {
      "post": {
        "description": "Runs the requested audit modules synchronously, writes findings via the deduplicated insert helper, and returns a per-category count plus the run id for log correlation.",
        "operationId": "run_audit_tenants__tenant_id__audit_run_post",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute audit categories and persist findings",
        "tags": [
          "audit"
        ]
      }
    },
...
    "/tenants/{tenant_id}/findings/{finding_id}/narrative": {
      "get": {
        "operationId": "get_narrative_tenants__tenant_id__findings__finding_id__narrative_get",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NarrativeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read or generate a finding narrative",
        "tags": [
          "narratives"
        ]
      }
    },
    "/tenants/{tenant_id}/findings/{finding_id}/narrative/regenerate": {
      "post": {
        "description": "Bypasses the cache and replaces any existing entry on success. Restricted to dev callers because it spends real money on every invocation.",
        "operationId": "regenerate_narrative_tenants__tenant_id__findings__finding_id__narrative_regenerate_post",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NarrativeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Force-regenerate a narrative (dev key only)",
        "tags": [
          "narratives"
        ]
      }
    },
...
  },
...
}